change type of vars in srcs to suppress error msgs

+++ src/BYTETracker.cpp
@@ -42,7 +42,7 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
 
     if (objects.size() > 0)
     {
-        for (int i = 0; i < objects.size(); i++)
+        for (size_t i = 0; i < objects.size(); i++)
         {
             std::vector<float> tlbr_;
             tlbr_.resize(4);
@@ -67,7 +67,7 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
     }
 
     // Add newly detected tracklets to tracked_stracks
-    for (int i = 0; i < this->tracked_stracks.size(); i++)
+    for (size_t i = 0; i < this->tracked_stracks.size(); i++)
     {
         if (!this->tracked_stracks[i].is_activated)
             unconfirmed.push_back(&this->tracked_stracks[i]);
@@ -87,7 +87,7 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
     std::vector<int> u_track, u_detection;
     linear_assignment(dists, dist_size, dist_size_size, match_thresh, matches, u_track, u_detection);
 
-    for (int i = 0; i < matches.size(); i++)
+    for (size_t i = 0; i < matches.size(); i++)
     {
         STrack *track = strack_pool[matches[i][0]];
         STrack *det = &detections[matches[i][1]];
@@ -104,14 +104,14 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
     }
 
     ////////////////// Step 3: Second association, using low score dets //////////////////
-    for (int i = 0; i < u_detection.size(); i++)
+    for (size_t i = 0; i < u_detection.size(); i++)
     {
         detections_cp.push_back(detections[u_detection[i]]);
     }
     detections.clear();
     detections.assign(detections_low.begin(), detections_low.end());
     
-    for (int i = 0; i < u_track.size(); i++)
+    for (size_t i = 0; i < u_track.size(); i++)
     {
         if (strack_pool[u_track[i]]->state == TrackState::Tracked)
         {
@@ -127,7 +127,7 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
     u_detection.clear();
     linear_assignment(dists, dist_size, dist_size_size, 0.5, matches, u_track, u_detection);
 
-    for (int i = 0; i < matches.size(); i++)
+    for (size_t i = 0; i < matches.size(); i++)
     {
         STrack *track = r_tracked_stracks[matches[i][0]];
         STrack *det = &detections[matches[i][1]];
@@ -143,7 +143,7 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
         }
     }
 
-    for (int i = 0; i < u_track.size(); i++)
+    for (size_t i = 0; i < u_track.size(); i++)
     {
         STrack *track = r_tracked_stracks[u_track[i]];
         if (track->state != TrackState::Lost)
@@ -165,13 +165,13 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
     u_detection.clear();
     linear_assignment(dists, dist_size, dist_size_size, 0.7, matches, u_unconfirmed, u_detection);
 
-    for (int i = 0; i < matches.size(); i++)
+    for (size_t i = 0; i < matches.size(); i++)
     {
         unconfirmed[matches[i][0]]->update(detections[matches[i][1]], this->frame_id);
         activated_stracks.push_back(*unconfirmed[matches[i][0]]);
     }
 
-    for (int i = 0; i < u_unconfirmed.size(); i++)
+    for (size_t i = 0; i < u_unconfirmed.size(); i++)
     {
         STrack *track = unconfirmed[u_unconfirmed[i]];
         track->mark_removed();
@@ -179,7 +179,7 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
     }
 
     ////////////////// Step 4: Init new stracks //////////////////
-    for (int i = 0; i < u_detection.size(); i++)
+    for (size_t i = 0; i < u_detection.size(); i++)
     {
         STrack *track = &detections[u_detection[i]];
         if (track->score < this->high_thresh)
@@ -189,7 +189,7 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
     }
 
     ////////////////// Step 5: Update state //////////////////
-    for (int i = 0; i < this->lost_stracks.size(); i++)
+    for (size_t i = 0; i < this->lost_stracks.size(); i++)
     {
         if (this->frame_id - this->lost_stracks[i].end_frame() > this->max_time_lost)
         {
@@ -198,7 +198,7 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
         }
     }
     
-    for (int i = 0; i < this->tracked_stracks.size(); i++)
+    for (size_t i = 0; i < this->tracked_stracks.size(); i++)
     {
         if (this->tracked_stracks[i].state == TrackState::Tracked)
         {
@@ -214,13 +214,13 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
     //std::cout << activated_stracks.size() << std::endl;
 
     this->lost_stracks = sub_stracks(this->lost_stracks, this->tracked_stracks);
-    for (int i = 0; i < lost_stracks.size(); i++)
+    for (size_t i = 0; i < lost_stracks.size(); i++)
     {
         this->lost_stracks.push_back(lost_stracks[i]);
     }
 
     this->lost_stracks = sub_stracks(this->lost_stracks, this->removed_stracks);
-    for (int i = 0; i < removed_stracks.size(); i++)
+    for (size_t i = 0; i < removed_stracks.size(); i++)
     {
         this->removed_stracks.push_back(removed_stracks[i]);
     }
@@ -232,7 +232,7 @@ std::vector<STrack> BYTETracker::update(const std::vector<Object>& objects)
     this->lost_stracks.clear();
     this->lost_stracks.assign(resb.begin(), resb.end());
     
-    for (int i = 0; i < this->tracked_stracks.size(); i++)
+    for (size_t i = 0; i < this->tracked_stracks.size(); i++)
     {
         if (this->tracked_stracks[i].is_activated)
         {
